src/statistics: fix warning report by VC2017
src/statistics.cpp(22): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

+++ src/statistics.cpp
@@ -18,8 +18,8 @@ engine::run()
 {
 	auto size = this->map.size();
 
-	int i = 1;
-	int median_index = size / 2;
+	size_t i = 1;
+	size_t median_index = size / 2;
 	if (size % 2)
 	{
 		median_index += 1;
